Time: 0 ms (100.00%) | Memory: 19.5 MB (94.03%) - LeetSync

+++ 662-maximum-width-of-binary-tree/maximum-width-of-binary-tree.cpp
@@ -12,14 +12,14 @@
 class Solution {
 public:
     int widthOfBinaryTree(TreeNode* root) {
-        queue<pair<TreeNode*, int>> q;
-        int ans=0;
+        queue<pair<TreeNode*, long long>> q;
+        long long ans=0;
         if(root==nullptr) return ans;
         q.push({root,0});
         while(!q.empty()){
             int s=q.size();
-            long min=q.front().second;
-            long st,last;
+            long long min=q.front().second;
+            long long st,last;
             for(int i=0;i<s;++i){
                 auto it=q.front();
                 q.pop();
@@ -30,7 +30,7 @@ public:
                 if(node->left) q.push({node->left, 2*num+1});
                 if(node->right) q.push({node->right, 2*num+2});
             }
-            ans=max(ans,static_cast<int>(last-st+1));
+            ans=max(ans,last-st+1);
         }
         return ans;
     }
